Stop translating display_name
Because display_name is used in other contexts, like Outlines,
(outside of normal template rendering) lazy translations lead to a
number of issues localizing/rendering the value.

Since this is just used to seed the default value first displayed to the
instructor in Studio [1], and doesn't affect LMS translation, there
shouldn't be a reason we need to try localizing at this level.

This follows the lead set by the ORA2 XBlock.

- [1] Studio only explicitly supports English.